Move Structer and enum to definition, else i wouldnt be able to use the class GrenadeInfo in "settings". Also improves readability
