+++ src/Hacks/grenadehelper.cpp
@@ -13,22 +13,22 @@ ImColor Settings::GrenadeHelper::infoMolotov = ImColor(100, 20, 20, 255);
 ImColor Settings::GrenadeHelper::infoSmoke = ImColor(30, 30, 30, 255);
 ImColor Settings::GrenadeHelper::infoFlash = ImColor(255, 255, 0, 255);
 
-std::vector<GrenadeHelper::GrenadeInfo> /**Settings::GrenadeHelper::**/grenadeInfos = {
-		GrenadeHelper::GrenadeInfo(GrenadeHelper::GrenadeType::SMOKE, Vector(1149.4f, -1183.97f, -141.51f), QAngle(-42.f, -165.f, 0.f), GrenadeHelper::ThrowType::NORMAL, pstring("Smoke A Stairs")),
-		GrenadeHelper::GrenadeInfo(GrenadeHelper::GrenadeType::FLASH, Vector(803, -1418, -44.91f), QAngle(-30.49f, -178.85f, 0.f), GrenadeHelper::ThrowType::NORMAL, pstring("Popflash A"))
+std::vector<GrenadeInfo> Settings::GrenadeHelper::grenadeInfos = {
+		GrenadeInfo(GrenadeType::SMOKE, Vector(1149.4f, -1183.97f, -141.51f), QAngle(-42.f, -165.f, 0.f), ThrowType::NORMAL, pstring("Smoke A Stairs")),
+		GrenadeInfo(GrenadeType::FLASH, Vector(803, -1418, -44.91f), QAngle(-30.49f, -178.85f, 0.f), ThrowType::NORMAL, pstring("Popflash A"))
 };
 
-bool GrenadeHelper::matches(C_BaseCombatWeapon* wpn, GrenadeHelper::GrenadeType type)
+bool GrenadeHelper::matches(C_BaseCombatWeapon* wpn, GrenadeType type)
 {
 	switch (type)
 	{
-		case GrenadeHelper::GrenadeType::HEGRENADE:
+		case GrenadeType::HEGRENADE:
 			return strcmp(wpn->GetCSWpnData()->szClassName, "weapon_hegrenade") == 0;
-		case GrenadeHelper::GrenadeType::SMOKE:
+		case GrenadeType::SMOKE:
 			return strcmp(wpn->GetCSWpnData()->szClassName, "weapon_smokegrenade") == 0;
-		case GrenadeHelper::GrenadeType::FLASH:
+		case GrenadeType::FLASH:
 			return strcmp(wpn->GetCSWpnData()->szClassName, "weapon_flashbang") == 0;
-		case GrenadeHelper::GrenadeType::MOLOTOV:
+		case GrenadeType::MOLOTOV:
 			return strcmp(wpn->GetCSWpnData()->szClassName, "weapon_molotov") == 0
 				   || strcmp(wpn->GetCSWpnData()->szClassName, "weapon_incgrenade") == 0;
 		default:
@@ -36,17 +36,17 @@ bool GrenadeHelper::matches(C_BaseCombatWeapon* wpn, GrenadeHelper::GrenadeType 
 	}
 }
 
-ImColor GrenadeHelper::getColor(GrenadeHelper::GrenadeType type)
+ImColor GrenadeHelper::getColor(GrenadeType type)
 {
 	switch (type)
 	{
-		case GrenadeHelper::GrenadeType::HEGRENADE:
+		case GrenadeType::HEGRENADE:
 			return Settings::GrenadeHelper::infoHE;
-		case GrenadeHelper::GrenadeType::SMOKE:
+		case GrenadeType::SMOKE:
 			return Settings::GrenadeHelper::infoSmoke;
-		case GrenadeHelper::GrenadeType::FLASH:
+		case GrenadeType::FLASH:
 			return Settings::GrenadeHelper::infoFlash;
-		case GrenadeHelper::GrenadeType::MOLOTOV:
+		case GrenadeType::MOLOTOV:
 			return Settings::GrenadeHelper::infoMolotov;
 		default:
 			return ImColor(255, 255, 255);
@@ -104,7 +104,7 @@ void GrenadeHelper::Paint()
 	if (!activeWeapon || activeWeapon->GetCSWpnData()->GetWeaponType() != CSWeaponType::WEAPONTYPE_GRENADE)
 		return;
 
-	for (auto grenadeInfo = grenadeInfos.begin(); grenadeInfo != grenadeInfos.end(); grenadeInfo++)
+	for (auto grenadeInfo = Settings::GrenadeHelper::grenadeInfos.begin(); grenadeInfo != Settings::GrenadeHelper::grenadeInfos.end(); grenadeInfo++)
 	{
 		if (Settings::GrenadeHelper::onlyMatchingInfos && !matches(activeWeapon,grenadeInfo->gType))
 			continue;
@@ -139,7 +139,7 @@ void GrenadeHelper::AimAssist(CUserCmd* cmd)
 	if (!shootThisTick && !shotLastTick)
 		return;
 
-	for (auto grenadeInfo = grenadeInfos.begin(); grenadeInfo != grenadeInfos.end(); grenadeInfo++)
+	for (auto grenadeInfo = Settings::GrenadeHelper::grenadeInfos.begin(); grenadeInfo != Settings::GrenadeHelper::grenadeInfos.end(); grenadeInfo++)
 	{
 		if (Settings::GrenadeHelper::onlyMatchingInfos && !matches(activeWeapon,grenadeInfo->gType))
 			continue;
@@ -183,7 +183,7 @@ void GrenadeHelper::CheckForUpdate()
 		//TODO Load into infos
 		return;
 	}
-	grenadeInfos = {};
+	Settings::GrenadeHelper::grenadeInfos = {};
 	cvar->ConsoleDPrintf("No Smokes for this map found.\n");
 }
